+++ .gitignore
@@ -53,10 +53,7 @@ x64/*
 download?source=files
 main.cpp
 */threads*
-<<<<<<< HEAD
 a
-=======
 
 #boost
 boost_1_62_0/*
->>>>>>> dev-SendRecieve
